+++ apps/core/src/common/guards/auth.guard.ts
@@ -27,8 +27,24 @@ export class AuthGuard implements CanActivate {
 
     const query = request.query as any
     const headers = request.headers
+
+    const session = await this.authService.getSessionUser(request.raw)
+
     const Authorization: string =
       headers.authorization || headers.Authorization || query.token
+
+    if (session) {
+      const isOwner = !!session.user.isOwner
+
+      if (isOwner) {
+        this.attachUserAndToken(
+          request,
+          await this.userService.getMaster(),
+          Authorization,
+        )
+        return true
+      }
+    }
 
     if (!Authorization) {
       throw new UnauthorizedException('未登录')
@@ -50,15 +66,9 @@ export class AuthGuard implements CanActivate {
     if (!isJWT(jwt)) {
       throw new UnauthorizedException('令牌无效')
     }
-    const ownVerifyPass = await this.authService.jwtServicePublic.verify(jwt)
+    const valid = await this.authService.jwtServicePublic.verify(jwt)
 
-    if (!ownVerifyPass) {
-      const cleckVerifyPass = await this.authService.verifyClerkJWT(jwt)
-      if (!cleckVerifyPass) {
-        throw new UnauthorizedException('身份过期')
-      }
-    }
-
+    if (!valid) throw new UnauthorizedException('身份过期')
     this.attachUserAndToken(
       request,
       await this.userService.getMaster(),