+++ apps/core/src/modules/aggregate/aggregate.service.ts
@@ -240,9 +240,7 @@ export class AggregateService {
         ),
 
       this.postService.model
-        .find({
-          hide: false,
-        })
+        .find()
         .populate('category')
         .then((list) =>
           list.map((doc) => {
@@ -286,7 +284,7 @@ export class AggregateService {
 
     const [posts, notes] = await Promise.all([
       this.postService.model
-        .find({ hide: false })
+        .find()
         .limit(10)
         .sort({ created: -1 })
         .populate('category'),
@@ -333,6 +331,7 @@ export class AggregateService {
         created: post.created!,
         modified: post.modified,
         link: baseURL + this.urlService.build(post),
+        images: post.images || [],
       }
     })
     const notesRss: RSSProps['data'] = notes.map((note) => {
@@ -343,6 +342,7 @@ export class AggregateService {
         created: note.created!,
         modified: note.modified,
         link: baseURL + this.urlService.build(note),
+        images: note.images || [],
       }
     })