+++ apps/core/src/modules/auth/auth.controller.ts
@@ -12,8 +12,10 @@ import {
   Delete,
   Get,
   NotFoundException,
+  Patch,
   Post,
   Query,
+  Req,
 } from '@nestjs/common'
 import { EventEmitter2 } from '@nestjs/event-emitter'
 
@@ -21,6 +23,7 @@ import { ApiController } from '~/common/decorators/api-controller.decorator'
 import { Auth } from '~/common/decorators/auth.decorator'
 import { EventBusEvents } from '~/constants/event-bus.constant'
 import { MongoIdDto } from '~/shared/dto/id.dto'
+import { FastifyBizRequest } from '~/transformers/get-req.transformer'
 
 import { AuthService } from './auth.service'
 
@@ -96,5 +99,29 @@ export class AuthController {
 
     this.eventEmitter.emit(EventBusEvents.TokenExpired, token)
     return 'OK'
+  }
+
+  @Patch('as-owner')
+  @Auth()
+  async oauthAsOwner() {
+    return this.authService.setCurrentOauthAsOwner()
+  }
+
+  @Get('session')
+  async getSession(@Req() req: FastifyBizRequest) {
+    const session = await this.authService.getSessionUser(req.raw)
+
+    if (!session) {
+      return null
+    }
+
+    const account = await this.authService.getOauthUserAccount(
+      session.providerAccountId,
+    )
+
+    return {
+      ...session.user,
+      ...account,
+    }
   }
 }