+++ apps/core/src/modules/auth/auth.interface.ts
@@ -1,0 +1,13 @@
+import type { AdapterSession, AdapterUser } from '@mx-space/complied/auth'
+
+export type SessionUser = AdapterSession & {
+  user: AdapterUser
+  provider: string
+  providerAccountId: string
+}
+
+declare module '@mx-space/complied/auth' {
+  export interface AdapterUser {
+    isOwner: boolean
+  }
+}