+++ apps/core/src/modules/backup/backup.service.ts
@@ -97,11 +97,9 @@ export class BackupService {
     // 用时间格式命名文件夹
     const dateDir = getMediumDateTime(new Date())
 
-    // TODO windows下使用2024-07-08_14:13:53路径会报错
     const backupDirPath = join(BACKUP_DIR, dateDir)
     mkdirp.sync(backupDirPath)
     try {
-      // TODO 需要装mongo-tools
       await $`mongodump --uri ${MONGO_DB.customConnectionString || MONGO_DB.uri} -d ${
         MONGO_DB.dbName
       }  ${flatten(