+++ apps/core/src/modules/configs/configs.default.ts
@@ -9,10 +9,10 @@ export const generateDefaultConfig: () => IConfig = () => ({
     keywords: [],
   },
   url: {
-    wsUrl: 'http://127.0.0.1:2333', // todo
-    adminUrl: 'http://127.0.0.1:2333/proxy/qaqdmin',
-    serverUrl: 'http://127.0.0.1:2333',
-    webUrl: 'http://127.0.0.1:2323',
+    wsUrl: 'http://localhost:2333', // todo
+    adminUrl: 'http://localhost:2333/proxy/qaqdmin',
+    serverUrl: 'http://localhost:2333',
+    webUrl: 'http://localhost:2323',
   },
   mailOptions: {
     enable: false,
@@ -68,12 +68,7 @@ export const generateDefaultConfig: () => IConfig = () => ({
     xLogSiteId: '',
     githubToken: '',
   },
-  clerkOptions: {
-    enable: false,
-    adminUserId: '',
-    pemKey: '',
-    secretKey: '',
-  },
+
   authSecurity: {
     disablePasswordLogin: false,
   },
@@ -83,5 +78,11 @@ export const generateDefaultConfig: () => IConfig = () => ({
     openAiEndpoint: '',
     openAiPreferredModel: 'gpt-3.5-turbo',
     openAiKey: '',
+    aiSummaryTargetLanguage: 'auto',
+  },
+  oauth: {
+    providers: [],
+    secrets: {},
+    public: {},
   },
 })