+++ apps/core/src/modules/server-time/server-time.middleware.ts
@@ -12,9 +12,13 @@ export async function trackResponseTimeMiddleware(
 
   res.setHeader('Content-Type', 'application/json')
   // cors
-  res.setHeader('Access-Control-Allow-Origin', '*')
+  res.setHeader(
+    'Access-Control-Allow-Origin',
+    req.headers.origin || req.headers.referer || req.headers.host || '*',
+  )
   res.setHeader('Access-Control-Allow-Methods', 'GET, POST, OPTIONS')
   res.setHeader('Access-Control-Allow-Headers', 'Content-Type')
+  res.setHeader('Access-Control-Allow-Credentials', 'true')
   res.setHeader('Access-Control-Max-Age', '86400')
   await next()